requirements.txt
flask==3.0.3
gunicorn==22.0.0
groq==0.9.0
edge-tts==6.1.12
Pillow==10.4.0
